Add files via upload
Here I have to extract the data from 100 URL it should only contains the article title and its content. But 2 URL was not able to extract the data as the was not available. As, from 98 URL I have extracted the data using response, beautiful soup and pandas libraries and saved it in csv file.